+++ iOS/banchango/Sources/Presentation/RecipeRecommend/View/RecipeRecommendView.swift
@@ -15,8 +15,6 @@ struct RecipeRecommendView: View {
         recipeRecommendVM.getRecipeRecommend()
     }
     
-    let colors:[Color] = [.purple, .pink, .orange]
-    
     var body: some View {
         GeometryReader { proxy in
             ScrollView(.vertical) {
@@ -29,43 +27,6 @@ struct RecipeRecommendView: View {
             }
         }.onAppear() {
             UIScrollView.appearance().isPagingEnabled = true
-        }
-    }
-}
-
-
-struct RecipeCardView: View {
-    let recipe: Recipe
-    
-    var body: some View {
-        ZStack {
-            RoundedRectangle(cornerRadius: 20)
-                .fill(Color.clear)
-                .background(
-                    AsyncImage(url: URL(string: recipe.image)) { image in
-                        image
-                            .resizable()
-                            .scaledToFill()
-                            .frame(width: 300, height: 550)
-                    } placeholder: {
-                        Color.gray.opacity(0.3)
-                    }
-                    .clipShape(RoundedRectangle(cornerRadius: 20))
-                )
-                .frame(width: 300, height: 550)
-                .overlay(
-                    VStack {
-                        Text(recipe.name)
-                            .font(.largeTitle)
-                            .foregroundColor(.black)
-                        Spacer()
-                        Text(recipe.introduction)
-                            .font(.caption)
-                            .foregroundColor(.gray)
-                            .multilineTextAlignment(.center)
-                    }
-                    .padding()
-                )
         }
     }
 }
